Consolidates discovery of Prefect server `Service`s to the base class
Following on the work in #16898 to unify Prefect's `Service`
implementations, this moves the discovery routine from the CLI to the
base `Service` implementation.  This also includes a couple of
additional niceties:

- Each service is now responsible for knowing which `Setting` enables
  it.  This eventually means that when adding a new service, we won't
  need to modify the `api/server.py` lifespan or the CLI.  The
  conditional toggling of services on and off will be handled by asking
  the service if it should be enabled.

- The way service names are handles is more uniform across all the
  subclasses.  I also cleaned up the descriptions of all services to
  make them more punchy and imperative.

- The table of services uses the setting name as part of how it shows
  whether the service is enabled, which means folks shouldn't have to
  hunt through the code or docs to find the setting.

We do still have an explicit listing of all the known service modules,
because I wasn't too keen to go down the rabbithole of automatic
discovery from entrypoints or anything like that.  Definitely open to
ideas on a better place for that.

This version doesn't unify how the `api/server.py` lifespan starts
services, but that will be coming in a future update.
