+++ src/prefect/server/services/cancellation_cleanup.py
@@ -11,10 +11,11 @@ import sqlalchemy as sa
 from sqlalchemy.sql.expression import or_
 
 import prefect.server.models as models
+import prefect.settings
 from prefect.server.database import PrefectDBInterface, orm_models
 from prefect.server.database.dependencies import db_injector
 from prefect.server.schemas import filters, states
-from prefect.server.services.loop_service import LoopService
+from prefect.server.services.base import LoopService
 from prefect.settings import PREFECT_API_SERVICES_CANCELLATION_CLEANUP_LOOP_SECONDS
 
 NON_TERMINAL_STATES = list(set(states.StateType) - states.TERMINAL_STATES)
@@ -22,9 +23,12 @@ NON_TERMINAL_STATES = list(set(states.StateType) - states.TERMINAL_STATES)
 
 class CancellationCleanup(LoopService):
     """
-    A simple loop service responsible for cancelling tasks and subflows left over from
-    cancelling flow runs
+    Cancels tasks and subflows of flow runs that have been cancelled
     """
+
+    @classmethod
+    def enabled_setting(cls) -> prefect.settings.Setting:
+        return prefect.settings.PREFECT_API_SERVICES_CANCELLATION_CLEANUP_ENABLED
 
     def __init__(self, loop_seconds: Optional[float] = None, **kwargs: Any):
         super().__init__(
